Stats 方法从 JSON 字符串改为 UniFFI 类型化 DTO
Rust:
- UDL: 新增 DateRangeDto/WritingStatsSummaryDto/ProjectStatsRecordDto/
  ProjectStatsSummaryDto/ChapterStatsRecordDto/ChapterStatsSummaryDto/
  DeviceStatsRecordDto/DeviceStatsSummaryDto/SpeedCurvePointDto/
  SpeedCurveSummaryDto/PlatformDto 定义
- UDL: 5 个 _json 方法替换为类型化方法
- app_service.rs: 对应方法返回 DTO 而非 String
- writing_stats/api.rs: days_count cast as u32

Android:
- AppServiceBridge: 5 个方法从 BridgeResult<String> 改为返回类型化模型
- StatsBridge: 去掉 parseJsonResult，直接使用类型化结果
- 新增 DateRangeDto/WritingStatsSummaryDto 等 UniFFI DTO → Kotlin 模型转换
- 重新生成 UniFFI Kotlin 绑定

MindMap/StarMap/Actions 因类型复杂（serde_json::Value、标签枚举、
#[serde(other)]）仍使用 JSON 字符串 + parseJsonResult 模式

+++ core/writer_core/src/api/types.rs
@@ -2018,7 +2018,7 @@ pub struct WritingStatsSummaryDto {
     pub total_net_delta_chars: i64,
     pub total_active_seconds: u64,
     pub total_sessions: u32,
-    pub days_count: usize,
+    pub days_count: u32,
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq)]
